Add PHP functions_framework buildpack
It installs the functions framework [1] if required and sets the web process to invoke
the function.

Also use the same flags for all invocations of `composer install`. php.ComposerInstall
took a slice of flags, but they were always the same and there are more and more copies.
Define them once in php.ComposerInstall itself instead.

[1] https://github.com/GoogleCloudPlatform/functions-framework-php

PiperOrigin-RevId: 307654147
Change-Id: Ibeec7e128cfaa048703a7c88a4ee97d9fb8c5fc8