Add a JVM flag to Maven builds that removes the `illegal reflective access` warning. This warning comes from [this Guice bug](https://github.com/google/guice/issues/1133) which is not going to be fixed soon in a version of `mvn` that we can use.
Modify the acceptance test so that most test cases will fail if they see the string `WARNING` in the build output. Adjust some of the test projects to remove some warnings that would have triggered this.

This is the second attempt at this fix. The first had to be rolled back because it didn't work with Java 8.

PiperOrigin-RevId: 317207989
Change-Id: If0732f2663de6b339525db0e197c554620b2fd60