In failure acceptance tests, ensure that the `$BUILDER_OUTPUT` file contains failure messages.
Currently we check the stdout and stderr of the build for the message, but we don't check whether the message also gets sent to `$BUILDER_OUTPUT`. If it doesn't, the result can be an undesirable experience when using Google Cloud Build: the error message does not get shown in the UI that launched the build, though typically it can be found by looking through the build logs.

As currently constructed, it's hard for an acceptance test to check this itself. The `$BUILDER_OUTPUT` logic happens inside the container where the build is taking place, and there's no easy way to extract the files it writes, at least when using `pack`. So instead, we define a new env var `$EXPECTED_BUILDER_OUTPUT` which is a regexp that the `$BUILDER_OUTPUT` text should contain. When a buildpack is writing `$BUILDER_OUTPUT`, if `$EXPECTED_BUILDER_OUTPUT` is also set then the buildpack logs a message saying whether or not the `$BUILDER_OUTPUT` text contained a match for the regexp. Then the acceptance test can check for a message saying that it did.

Also update the GCF php74 buildpack so it gets the error output from a failing `php -l` from stdout rather than stderr. Similar problems were already fixed for the GCF python38 and java11 buildpacks, and were the motivation for this change.

PiperOrigin-RevId: 318106870
Change-Id: Iae81f3f9b15270fc506e9020d26957459be184d7