Fix gradle devmode and acceptance test framework
* Mark the gradle layer as launch=true if devmode is enabled.
* Fail the test if we run out of retried when invoking the app after source modification.

The gradle buildpack was not set as launch=true, so when source code was updated, the application hung with `.devmode_rebuild.sh: gradle: No such file or directory`. The test framework did not correctly handle the case where we ran out of retries and the last request resulted in an error, causing the test to be reported as passing after 10 minutes.

PiperOrigin-RevId: 324105851
Change-Id: I7eee57a7179bc337378d21733a651b5175a3921a