Support pre-release Go versions
Pre-release Go versions have version strings ending in, for example, "rc1" or "beta2".

The Go buildpacks don't support this; they try to parse the version string as semver
and fail when the suffix is encountered. This commit modifies the regexp used to parse
`go version` output to remove the suffix. Handling of versions read from go.mod files
is unchanged because pre-release suffixes aren't included there.

PiperOrigin-RevId: 326576527
Change-Id: Iba04f4359a77b216c2d32c4e8d97684d593f3876