Prevent version conflicts between vendored directories
There is a subtle bug here that manifests as an obscure error when the metadata is fetched from the context. The previous implementation had vendor directories in both the main app and in the function source code. In the case where the function depended on cloud.google.com/go/functions/metadata (as most event-style functions do), there would be two copies of cloud.google.com/go/functions metadata, one in serverless_function_app/vendor and one in <function>/vendor. The framework called from the main function would use the former to set the event metadata on the context, and the function would use the latter to retrieve it. This caused an opaque error where the compiler would identify the keys as different objects, and therefore on the function's attempt to retrieve it would return a "not-found" error.

This change copies the vendor directory directly into GOPATH, rather than into separate vendored directories, which resolves the multiple-versions problem.

PiperOrigin-RevId: 327131979
Change-Id: I6f9edd7b0462d1d35504bae7feec96cedba14b68