Add tests for vendored function with framework in gcf/go113
Adds a new Setup function to the acceptance test config struct that can be used to modify a copy of the app source code before running the tests. In this case, the setup functions run `go mod vendor` using the builder under test to download the modules specified in `go.mod` and then remove the `go.mod` itself.

PiperOrigin-RevId: 327677379
Change-Id: I3734440d56031904f002eeb3364daccaa0438bea