Roll back change that addressed version conflicts between vendor directories
Commit rolled back:
https://github.com/GoogleCloudPlatform/buildpacks/commit/6879fdb359be84f888b6e0f2fc9506426dff4651

That change fixed one issue with vendor directories that don't include the
functions framework, but it introduced another:

If the function...
- provides a vendor dir excluding the FF
- and the vendor dir includes a package with the same parent as a dep of the FF
- and at least one FF dep under that parent package is not in the vendor dir

...then the build fails while `go get`ting the framework (`go get` finds the parent
package of the dependency but does not find a .git directory inside it).

Fixing the new issue would lead to more hard-to-debug errors with version
conflicts, so revert the change instead.

Also log a warning message to inform the user that providing a vendor dir
that does not contain the functions framework may lead to unexpected errors.

PiperOrigin-RevId: 336180624
Change-Id: I50f0c226a19c5216e8fd4e2f52dfb061c8999582