Use `go mod vendor` to fetch FF in GOPATH case so that all deps' versions are pinned
Previously we used `go get` to install the functions framework and its dependencies.
After installing we did a `git checkout` to get the desired version of the FF. The
versions of the FF's dependencies were not changed however: they were left at HEAD,
as such is the behavior of `go get`.

This means that a bad commit in any dependency, or one that interacts poorly with
the FF itself, could cause errors. The workarounds are to (1) manually vendor the
FF or (2) switch to a go.mod build.

Instead use `go mod vendor` because that allows the versions of all of the
framework's dependencies to be pinned as specified in the framework's go.mod.
Make a module in a temp dir, add a dependency on the FF, download all deps with
`go mod vendor`, and then copy the contents of the vendor dir into GOPATH.

PiperOrigin-RevId: 339121939
Change-Id: Id58d6d1f11085fee28f2e7e5818e523a5a27b5e7