Use a single vendor dir in the case that the FF is vendored
The current state is that if a vendor directory including the functions framework is
provided, it is copied from the function package to the generated main package.
This results in two copies of the dependencies that, while identical, are not
treated as such by Go due to their different locations.

There's some precedent in addressing issues with vendoring.

https://github.com/GoogleCloudPlatform/buildpacks/commit/6879fdb359be84f888b6e0f2fc9506426dff4651 attempted to address this very issue of two copies of the same dependency.

https://github.com/GoogleCloudPlatform/buildpacks/commit/676a49b81a0756866b76014bf19e2f2442156b12#diff-83397f049f56b454132e0a595161ba477f50d8c95cb38fa1326eb15f714c03b2 rolled back that
commit because along with fixing one issue it introduced an edge case related to
installing the functions framework on top of a vendor dir that doesn't include it.

This commit addresses only the case where the functions framework is included in the
vendor dir, i.e., it includes everything required to build, and we do not install
the framework on behalf of the user. In that case there's no need to use two copies
of the vendor dir and thereby cause confusing type inequality. Instead we'll move
the vendor dir to the directory that contains both the function package and the
generated main package, allowing both to use the same set of deps.

PiperOrigin-RevId: 355478941
Change-Id: Ic48f5f54c8badb1af6c616d209d046af4aa50573