Add support for GOOGLE_INTERNAL_BUILD_DIR in go.gomod buildpack
In the go.appengine_gomod buildpack, the application source is moved to a layer so running `go mod tidy` has no effect on the source that is used the compile the application. This also makes existing acceptance tests pass for Go 1.16, which requires go.sum to be present before `go mod download`.

PiperOrigin-RevId: 359633154
Change-Id: Ib438ceb12c1eeb6b6ea2804052943afd17e93e93