fix: correctly detect vcpkg version changes
vcpkg has two separate versions, the tool and the ports. We were only checking the version of the tool, but unfortunately the current release has a minor bug and reports an old version. We were not checking the version of the ports, which we now
do by checking the SHA256 of the baseline file.

PiperOrigin-RevId: 364376608
Change-Id: I2a0fae74806930ee166b53bd3c5994a0455d2289