Allow version pinning java functions framework
Currently, the FF java buildpack always dowloads the latest invoker jar
and uses it to bootstrap the user's function. This commit updates the
buildpack to check if the invoker is listed as a maven or gradle dependency
and if so, use the downloaded dependency instead of downloading the latest
version.

PiperOrigin-RevId: 375539005
Change-Id: I94144db4fd4fdc9697c0fa3ed17f85a5809b14b6