Create a buildpack and builder for the nodejs8 GCF runtime
This commit creates a new buildpack that builds nodejs GCF applications using the nodejs8 GCF worker.

Most of the existing nodejs acceptance tests are not applicable for this builder because both Firebase and the Functions Framework block use of node versions < 10. I will add some new acceptance specific to this builder in a subsequent commit.

PiperOrigin-RevId: 384397683
Change-Id: I69d426e2f90a42ca1fa23bf9c1d702c3ca9cd293