Fix support for functions with vendored dependencies to Go 1.14+
The current implementation has a bug which causes all dependencies to be re-downloaded when running `go mod tidy`. Functions with `vendor` do not require a `go.sum` file, which makes `go mod tidy` unnecessary. However, they require the `vendor` directory to be at root.

Just moving the `vendor` directory to the application root is not enough, as it must be consistent with the root `go.mod`. We can copy the functions `go.mod` to the root, but keeping the current `require/replace` logic to import the function then also fails, because `vendor` will again be inconsistent with `go.mod`.

To resolve this:

1. Require that the functions framework is included in `vendor`. Downloading it during the conversion step is likely to cause version conflicts with function-declared dependencies as the framework uses popular packages. We cannot simply require it and re-run `go mod vendor`, as that will cause all modules to be re-downloaded, defeating the point of `vendor`.
2. Create `main.go` in functions vendor directory with a module name `functions.local/app`. There, require the user function using its full module name (instead of `replace`).
3. Build the app by referring to the full package name, i.e. `go build functions.local/app`. Go will correctly find the module in `vendor`.

PiperOrigin-RevId: 389055628
Change-Id: I489193a996282868ec3ed309941d0ec916d3e311