Fix Go 1.11 legacy buildpack module path errors
The legacy_worker (and functions_framework) buildpack currently enforce that user functions that include a go.mod must use a module name that includes a '.' in the path. This is a requirement of Golang when modules need to be downloaded. However, for local-only modules that are referred to by source directory, [dot-less module names can be used](https://golang.org/ref/mod#go-mod-file-ident). Since we have the user function source code locally at build time, we do not need to download the user function module when building the main application's Golang module with the GCF buildpacks, so we do not need to enforce users use dot-less module paths.

Unfortunately, until Go 1.14, the Golang command tool enforced dot-less module paths in all cases when running `go mod edit -require <module name>`, which is how we build the main application's `go.mod` file. To make the Go 1.11 legacy_worker buildpack work, we need to make the `go.mod` file ourselves to avoid the error.

PiperOrigin-RevId: 401056598
Change-Id: Id8ca1a7e12876cf2ef358114c8c05dd6c79cb58a