Update Go Functions Framework default to v1.5.0
Also, alter functions_framework buildpack to be able to handle declarative function signatures. This will write out a different main that does NOT require the FUNCTION_TARGET to be an exported function on the user's package. When the user's function package imports "github.com/GoogleCloudPlatform/functions-framework-go/functions", it's assumed that declarative function signatures are being used.

PiperOrigin-RevId: 409282261
Change-Id: I59da457fdba72a4be284e477e6f651b47426b90f