Cleanup the missing entrypoint BP
This buildpack exists to break the build with a clear error message when
Python customers don't specify an entrypoint. Before this commit we were
including the python.runtime in the order.group so we could detect python
groups, but that would cause us to install Python before we failed the
build.

PiperOrigin-RevId: 421911470
Change-Id: If9cd628a9033f3a1eff23d65a02f4f9861e4d55a