Add a mechanism for retrieving the associated runtime version for a given SDK version via .NET's release-index.json
We are currently installing the runtime by attempting to delete the SDK from the SDK tarball and hope that the runtime remains. In a future commit, we will instead get the correct, associated version of the runtime for a given SDK version. This change will make that possible by enabling us to figure out the associated runtime version for a given SDK version.

The 'dotnet' command does not provide a documented mechanism for getting the runtime version for the SDK. It is possible to ascertain this value by grepping for certain strings but that method would requiring shelling out to dotnet and then trying to string match which seems more brittle that the release-index which is supported by .NET (the release-index is actually an object in the azure blob store, and we already have a dependency on the same bucket for getting the latest version).

PiperOrigin-RevId: 422892265
Change-Id: If818dea23cf0b401ace5551a8e83ac15577c3c3c