Merge the yarn-gcp-build buildpack into the yarn buildpack
This commit updates the logic of the yarn buildpack in a few ways:

1. Always install dependencies and devDependencies, then purge devDependencies in the final app image.
2. Use `--prefer-offline` when running `yarn install` to avoid re-downloading cached dependencies.
3. Merge the logic to run the `gcp-build` script into the yarn buildpack and delete the separate `yarn-gcp-build` buildpack.

This pattern of always installing then purging devDependencies is more idiomatic and makes it easier for developers who wish to use the `postinstall` lifecycle hook.

This refactoring is in preparation for an upcoming CL to add Yarn2 support.

PiperOrigin-RevId: 422953938
Change-Id: I92858d3b0977076d3cd3c5511d7700f748d229f9