Add support for Yarn2
This commit adds support for Yarn2 and its zero-install Plug'n'Play feature. The overall flow for Yarn2 is similar to Yarn1 but some of the CLI flags have changed. Additionally when a developer enables Plug'n'Play mode, we don't cache the node_modules in a layer, because they will be included with the application source.

Fixes #150, #107

PiperOrigin-RevId: 424395115
Change-Id: Ic3e88749e97ae8c8a4240965e2009be1a851bac8