Add a facility for getting the appropriate SDK release channel for a given Project
The purpose of this function is to figure out the appropriate SDK channel for which we should look for an SDK version.

A Project may have a "TargetFramework" field that targets a specific framework version. This is not the same as SDK version, as an SDK can support more than one framework version. For example, the latest 5.x SDK supports .NET core 1.0 -> 5.x. However, the 6.0 SDK only supports 6.0.

There is an additional bit of complexity, which is that while the 5.0 SDK may support building and running netcore 1.0 -> 3.1 apps, a given .NET runtime version can only run apps with the same major version. That is to say, the 3.1 rutnime can run 3.0 and 3.1 apps, but the 5.0 runtime cannot run 3.0 apps. For that reason we limit the SDK version to the greatest compatible version for which the runtime version is compatible. This is to take advantage of any security fixes that have been made.

PiperOrigin-RevId: 425968398
Change-Id: I6a612308fca7413eb421c3ac541f4d8c47673402