Fix AppEngine tests for unified builder
They were broken by moving the stack to a base image without go pre-installed.
In addition, a line that was erronously removed from common_test.go in a merge was added back.

The go 1.11 tests were previously passing because a newer version of go (1.16) was pre-installed on the build image. Once we moved to the new build image without the pre-installed go they started failing. The reason why they failed is many of the test applications had a go 1.13 requirement in their go.mod file. This requirement was not met because, for go 1.11, the installed version of go is 1.11. So we can keep using the same test cases across all supported versions of go the required version of go is downgraded to 1.11 for the affected apps.

PiperOrigin-RevId: 442902727
Change-Id: Ice797b0c9d00ee4fa076a6f4cfedbf509d0ba81e