Implement the .NET unified builder
The tests are copied & pasted from their other locations. They are not yet testing across multiple versions (other than the explicit version selection of a few of the GCP tests).
The GAE tests are disabled as they are not yet passing as they need a run image which has /serve on it.

PiperOrigin-RevId: 444361163
Change-Id: If8312b425889fee6faf6e88b50578fb20e14dfcf