Move the gcp single version test cases into the main Acceptance tests and move the gcf go111 test cases into a seperate blaze target
This will enable a future commit which will add blaze functionality which will run all tests in the rule across a number of go versions. Since the GCF go111 tests are only intended to run against a single version they need to be split off from the rest of the tests.

PiperOrigin-RevId: 444865838
Change-Id: I3b568de31f122901d92a043bb36576c7de6a0545