move npm upgrade logic to npm buildpack
This enables customers on GAE and GCF to configure the NPM version used in their applications. Previously it wasn't possible because the nodejs/runtime buildpack is not used for GCF and GAE.

I also updated some of the acceptance tests to validate the npm installed in the final application image.

PiperOrigin-RevId: 448162633
Change-Id: I4c19d893bd589ee753e3ece68db4d0db68d32ef0