Change acceptance.setupFunc to pass in a SetupContext rather than a list of arguments.
The next commit will add another argument to setupFunc. Doing so requires updating each implementation of setupFunc and increases the number of arguments to 3. As we continue to add more scenarios and arguments, the list will become cumbersome and the refactors will be toilsome.

By adding the SetupContext struct the refactor only needs to be done this time and in the future new arguments can be added without changing the function signatures.

PiperOrigin-RevId: 449345221
Change-Id: If5b74b2cc295ec6f3791313abe105b0c6a0a6d45