Change the name of acceptance.CreateBuilder to acceptance.ProvisionImages; Return the name of the run image when overriden
To better match the responsibilites of CreateBuilder the function is renamed to Provision Images. The responsibilites are:
* Ensure that the appropriate tools are present on the system (pack, etc)
* Retag the builder image if builderName is specified
* Create the builder image from the builder.toml if builderName is not specified
* Pull the run and build images
* Pull the override run images

The run image is returned to reduce the number of places in acceptance.go that need to reason about the runImageOverride parameter. In addition, the next commit in this chain dynamically creates a new run image with new labels. For this commit, we need TestApp(...) to have access to the new run image name. Rather than rely on global state, return the new run image name so that it can be passed in.

PiperOrigin-RevId: 449812389
Change-Id: I14518c933b066e6c47c3bdf2ca1b14cbf1ce93d5