fix license validation in Ubuntu 22 run images
Before we publish images we validate that all the correct license information exists. This validation is based on "licenses.yaml" files published in a specific directory. This commit moves the licenses.yaml files into the right place in `google.gae.22` and creates an empty `licenses.yaml` for `google.min.22` because it has no licenses to validate.

PiperOrigin-RevId: 455699640
Change-Id: I85e2aa38a334699d30ad0aa7e4b17e5f8be73aff