Update dotnet logic to prioritize publish mechanism
When both a project file and a runtimeconfig.json are
present, prioritize using the project file, and publish.

If a project file is found and GOOGLE_ASP_NET_CORE_VERSION
is not set, use this information to determine both the
sdk and runtime versions.

Otherwise, the use runtimeconfig.json found in the upload root.

Also sets dotnet to publish to a layer instead of workspace.

PiperOrigin-RevId: 474409994
Change-Id: I9d614b27057f59edd01b3c9cfa0d6391f47120b2