fix dotnet SDK cache invalidation bug
We should never re-use cache layers when customers switch stacks on GCF. We had logic to check this in the runtime package, but the dotnet.sdk buildpack had logic to check if the cache was valid before calling InstallTarballIfNotCached. This commit, fixes the issue by inverting the logic in the SDK buildpack so instead of detecting when to keep the cache and return early, it detects when the cache should be cleared. This way we always call through to InstallTarballIfNotCached.

PiperOrigin-RevId: 486165056
Change-Id: Ie371ce795a184c056743f5287ac982aad2ab97a2