fix bug in python exec prefix logic
While testing python 3.11 I found a bug in the logic for parsing the exec prefix where the layer path was being computed as "/layers/google.python.runtime/python1". This commit changes it to use a regex and adds some unit tests.

With this CL all GCF and GCP acceptance tests pass using python 3.11.

PiperOrigin-RevId: 488000846
Change-Id: I9137aaa9a0cbc5c13d873a75ca55949affdc0694