update to buildpacks lifecycle v0.16.0
The lifecyle binary is included in all of our builders and invokes the buildpacks in accordance with the specification. This version includes support for new buildpacks features (eg SBoMs) and is compiled with a newer version of Go that patches some CVEs that show up in our base image scans.

PiperOrigin-RevId: 509941323
Change-Id: I58c3587a4f767e222148555c6e721cef33f22295