Fix buildpack metrics only being reported for the last buildpack that runs.
Because there is only one file for Cloud Build build output, all buildpacks in the builder must share and mutate the same file.

This fixes an issue with metrics reporting where the last buildpack was overwriting the previous buildpacks' metrics instead of adding to them.

Also, fix a similar issue for InstalledRuntimeVersions.

PiperOrigin-RevId: 526074283
Change-Id: Ife61807a56aba90b158dcf16d09ccffc1adebd4e