more robust entry point logic for nodejs
Currently we set the default entry point for nodejs applications to `npm run start` even if the package.json does not define a start script. This commit updates the logic to use fall back to either the "main" script from the package.json or use `node index.js` if neither is available.

fixes #375

PiperOrigin-RevId: 602032240
Change-Id: I9776c9f2f4dca514fb0f3c9dc6ba93ef4ab375bb